Logic for entities and more tests

+++ src/entities/Weapon.java
@@ -10,7 +10,7 @@ import interfaces.Entity;
 public abstract class Weapon extends Pickupable{
 	private int baseDamage;
 
-	public Weapon(String name, float x, float y, int width, int height, Type type, int damage) {
+	public Weapon(String name, double x, double y, int width, int height, Type type, int damage) {
 		super(name, x, y, width, height, type);
 		this.baseDamage = damage;
 	}
